Renamed store folder to redux
