+++ src/App.tsx
@@ -2,7 +2,8 @@ import React from "react";
 import { BrowserRouter as Router } from "react-router-dom";
 import { Provider } from "react-redux";
 import Pages from "pages";
-import store from "store";
+import store from "reduxUtils";
+
 import "index.css";
 
 function App() {
